Fix unit tests and tidy up
